feat: add match score display and improve entry count handling
This commit introduces a new match score display feature in the UI, enhancing user engagement with live sports updates. Key changes include:

- Added CSS styles for match score presentation, including live status indicators and tooltips.
- Updated JavaScript to fetch and render match scores dynamically based on the title.
- Refactored entry count handling to use a more robust parsing and formatting approach.
- Improved the overall structure of topic cards to accommodate match score information.

These enhancements aim to provide users with real-time updates and a more informative interface.

+++ src/actions/js/topics.js
@@ -87,7 +87,7 @@ export const Topics = {
     const newTopic = {
       ...topic,
       lastChecked: Date.now(),
-      lastEntryCount: parseInt(topic.entryCount)
+      lastEntryCount: parseFloat(topic.entryCount)
     };
 
     const result = await Storage.addFollowing(newTopic);
